added return confirm

+++ src/main/webapp/view.jsp
@@ -113,7 +113,7 @@
 				if(userID != null && userID.equals(tzs.getUserID())){
 			%>
 					<a href="update.jsp?tzsID=<%= tzsID %>" class="btn btn-primary">수정</a>
-					<a href="deleteAction.jsp?tzsID=<%= tzsID %>" class="btn btn-primary">삭제</a>
+					<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?tzsID=<%= tzsID %>" class="btn btn-primary">삭제</a>
 			<% 
 				}
 			%>
